try to refactor tests a bit
